\name{sagging}
\alias{sagging}
\docType{package}
\title{
This package includes several methods that can be used to predict the class of data.
}
\description{
sagging implements several methods for the classifier of  data.}
\details{
\tabular{ll}{
Package: \tab sagging\cr
Type: \tab Package\cr
Version: \tab 0.5\cr
Date: \tab 2016-11-23\cr
License: \tab GPL-2\cr
}
}
\author{
Maintainer: Saeid Amiri <saeid.amiri1@gmail.com>
}
\references{
Amiri, S. (2016). Classification via Aggregating on permutated variables.

\url{https://sites.google.com/site/saeidamiri1/codes/sagging}

}
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%#~~ simple examples of the most important functions ~~
%}
